Implement minimal viable exporter
Reads from the firehose, writes to promtail.

One thing that's slightly annoying is that Loki wants the logs to arrive
in order. Since the logs that come out of the firehose are not ordered
we're pushing them into Loki with the timestamp set as `time.Now()` as
the exporter sees it.

This isn't great, because:

1) The timestamps on the logs won't be the time they were actually
raised
2) We can't run multiple exporters writing to the same Loki, as they'll
end up racing each other

We might be able to resolve the first issue by doing some buffering.